booking-test/app/requirements.txt
aiogram
python-dotenv
pytz
psycopg2-binary
